requirements.txt
# Ensure you have the 'mcp.py' library available.
# It might be provided or need a specific install method.
# If it's a standard package, add its name here.

Flask             # For the custom_weather_api.py
requests
python-dotenv
PyGithub       # Used by the mcp.py example, so likely needed
mcp>=0.1.0
pymupdf>=1.21.1
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10
langchain>=0.0.335
langsmith>=0.0.65
faiss-cpu>=1.10.0
flask-cors
